added simple npc logic

+++ fish_game/arkiologic_game/UI.cpp
@@ -2,11 +2,11 @@
 
 
 void UI::draw(std::vector<CHAR_INFO>& buffer, Screen& screen) {
-	x = poses_arr[screen_pos][0] * screen.cols;
-	x -= wighth * sgn(x);
+	x = float(poses_arr[screen_pos][0] * screen.cols);
+	x -= wighth * float(sgn(x));
 
-	y = poses_arr[screen_pos][1] * screen.rows / 2;
-	y -= (hight + 1) * sgn(y);
+	y = float(poses_arr[screen_pos][1] * screen.rows / 2);
+	y -= (hight + 1) * float(sgn(y));
 
 	TextSquere::draw(buffer, screen);
 }
